pyproject.toml
[project]
name = "qa-index-worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.31",
    "langchain-huggingface>=0.3.1",
    "pydantic-settings>=2.11.0",
    "pyyaml>=6.0.3",
    "langchain-unstructured>=0.1.6",
    "httpx>=0.28.1",
    "pypdf>=6.1.2",
    "rapidocr-onnxruntime>=1.4.4",
    "pytesseract>=0.3.13",
]
